Add contributors, files, currentUserPermissions, and files fields to the Preprint model.
- Consolidate contributor-related methods to a ContributorMixin to be shared between the node and preprint models.
- Add a preprint relationship to the contributor model - tentatively trying to share this between nodes and preprints?
- Update contributor adapter so requests can be made to a preprint or a node.
