+++ addon/models/preprint.js
@@ -1,6 +1,7 @@
 import Ember from 'ember';
 import DS from 'ember-data';
 import OsfModel from './osf-model';
+import ContributorMixin from 'ember-osf/mixins/contributor-mixin';
 
 /**
  * @module ember-osf
@@ -16,7 +17,7 @@ import OsfModel from './osf-model';
  * https://api.osf.io/v2/docs/#!/v2/User_Preprints_GET
  * @class Preprint
  */
-export default OsfModel.extend({
+export default OsfModel.extend(ContributorMixin, {
 
     title: DS.attr('fixstring'),
     // TODO: May be a relationship in the future pending APIv2 changes
@@ -34,15 +35,22 @@ export default OsfModel.extend({
     preprintDoiCreated: DS.attr('date'),
     description: DS.attr('fixstring'),
     tags: DS.attr(),
+    public: DS.attr('boolean'),
+    // List of strings
+    currentUserPermissions: DS.attr(),
 
     // Relationships
     node: DS.belongsTo('node', { inverse: null, async: true }),
     license: DS.belongsTo('license', { inverse: null }),
     primaryFile: DS.belongsTo('file', { inverse: null }),
     provider: DS.belongsTo('preprint-provider', { inverse: 'preprints', async: true }),
+    files: DS.hasMany('file-provider'),
     reviewActions: DS.hasMany('review-action', { inverse: 'target', async: true }),
-    contributors: DS.hasMany('contributors', { async: true }),
-
+    contributors: DS.hasMany('contributors', {
+        allowBulkUpdate: true,
+        allowBulkRemove: true,
+        inverse: 'preprint'
+    }),
     uniqueSubjects: Ember.computed('subjects', function() {
         if (!this.get('subjects')) return [];
         return this.get('subjects').reduce((acc, val) => acc.concat(val), []).uniqBy('id');
